review changes + unittest fixed
